Creating ndarray from irregular nested sequences issue resolved

+++ script/util/pose.py
@@ -1,11 +1,9 @@
 import numpy as np
 
 def skew_symmetric(v):
-    return np.array(
-        [[0, -v[2], v[1]],
-         [v[2], 0, -v[0]],
-         [-v[1], v[0], 0]]
-    )
+    return np.array([[0, -v[2], v[1]],
+                     [v[2], 0, -v[0]],
+                     [-v[1], v[0], 0]], dtype=float)
 
 class Quaternion:
